refactoring project tree
