+++ src/fair/tests/data/b2_api.json
@@ -1,0 +1,8 @@
+{
+  "b2access_registration_link": "https://b2access.eudat.eu/", 
+  "b2note_url": "https://b2note.bsc.es/interface_main.html", 
+  "site_function": "trng", 
+  "terms_of_use_link": "http://hdl.handle.net/11304/e43b2e3f-83c5-4e3f-b8b7-18d38d37a6cd", 
+  "training_site_link": "", 
+  "version": "2.1.0"
+}
