--- src/pocs/pyld/b2_community_EUDAT_e9b9792e-79fb-4b07-b6b4-b9c2bd06d095.json
@@ -1,25 +1,0 @@
-{
-  "created": "Wed, 21 Dec 2016 08:57:40 GMT", 
-  "description": "The big Eudat community. Use this community if no other is suited for you", 
-  "id": "e9b9792e-79fb-4b07-b6b4-b9c2bd06d095", 
-  "links": {
-    "self": "https://trng-b2share.eudat.eu/api/communities/e9b9792e-79fb-4b07-b6b4-b9c2bd06d095"
-  }, 
-  "logo": "/img/communities/eudat.png", 
-  "name": "EUDAT", 
-  "publication_workflow": "direct_publish", 
-  "restricted_submission": false, 
-  "roles": {
-    "admin": {
-      "description": "Admin role of the community \"EUDAT\"", 
-      "id": 11, 
-      "name": "com:e9b9792e79fb4b07b6b4b9c2bd06d095:admin"
-    }, 
-    "member": {
-      "description": "Member role of the community \"EUDAT\"", 
-      "id": 12, 
-      "name": "com:e9b9792e79fb4b07b6b4b9c2bd06d095:member"
-    }
-  }, 
-  "updated": "Wed, 21 Dec 2016 08:57:40 GMT"
-}
